AERN2.MP.Accuracy

Description

A type for roughly measuring the accuracy of an enclosure.

Documentation

data Accuracy

A non-negative Double value to serve as an error bound. Arithmetic is rounded towards +infinity.

Constructors

NoInformation

Exact

bits :: ConvertibleExactly t Accuracy => t -> Accuracy

fromAccuracy :: Accuracy -> Integer

class HasAccuracy a where

Minimal complete definition

getAccuracy

Methods

getAccuracy :: a -> Accuracy

class HasAccuracyGuide a where

Minimal complete definition

getAccuracyGuide

Methods

getAccuracyGuide :: a -> Accuracy

class HasAccuracyGuide a => CanSetAccuracyGuide a where

Minimal complete definition

setAccuracyGuide

Methods

setAccuracyGuide :: Accuracy -> a -> a

adjustAccuracyGuide :: CanSetAccuracyGuide a => (Accuracy -> Accuracy) -> a -> a

getFiniteAccuracy :: (HasAccuracy t, HasPrecision t) => t -> Accuracy

Return accuracy, except when the element is Exact, return its nominal Precision dressed as Accuracy. This function is useful when we have a convergent sequence where all elements happen to be actually equal to the limit and we need the property that the sequence elements keep improving.

ac2prec :: Accuracy -> Precision

class CanReduceSizeUsingAccuracyGuide t where

Minimal complete definition

reduceSizeUsingAccuracyGuide

Methods

reduceSizeUsingAccuracyGuide :: Accuracy -> t -> t

specCanReduceSizeUsingAccuracyGuide :: (CanReduceSizeUsingAccuracyGuide t, HasEqCertainly t t, Arbitrary t, Show t) => T t -> Spec

iterateUntilAccurate :: HasAccuracy t => Accuracy -> (Precision -> Maybe t) -> [(Precision, Maybe t)]

convergentList2CauchySeq :: HasAccuracy t => [t] -> Accuracy -> t

seqByPrecision2CauchySeq :: HasAccuracy t => (Precision -> t) -> Accuracy -> t

setPrecisionAtLeastAccuracy :: CanSetPrecision t => Accuracy -> t -> t

Change the precision so that it is at least as high as the supplied accuracy (assuming the accuracy is finite).

class HasApproximate t where

An unsafe approximation of an enclosure or exact value, useful mainly for showing something brief and readable to humans.

Minimal complete definition

getApproximate

Associated Types

type Approximate t

Methods

getApproximate :: Accuracy -> t -> Approximate t