AERN2.MP.Ball.Comparisons

Description

Comparisons of arbitrary precision dyadic balls

Auxiliary types

module AERN2.Norm

Ball operations (see also instances)

reducePrecionIfInaccurate :: MPBall -> MPBall

Reduce the precision of the ball centre if the accuracy of the ball is poor.

More precisely, reduce the precision of the centre so that the ulp is approximately (radius / 1024), unless the ulp is already lower than this.

intersect :: MPBall -> MPBall -> MPBall

Helpers for constructing ball functions

byEndpointsMP :: (MPFloat -> MPFloat -> MPFloat) -> MPBall -> MPBall -> MPBall

Computes an *increasing* ball fucntion f from *exact* MPFR operations.

Orphan instances

CanTestPosNeg MPBall

Methods

isCertainlyPositive :: MPBall -> Bool

isCertainlyNonNegative :: MPBall -> Bool

isCertainlyNegative :: MPBall -> Bool

isCertainlyNonPositive :: MPBall -> Bool

CanTestInteger MPBall

Methods

certainlyNotInteger :: MPBall -> Bool

certainlyInteger :: MPBall -> Bool

certainlyIntegerGetIt :: MPBall -> Maybe Integer

CanTestZero MPBall

Methods

isCertainlyZero :: MPBall -> Bool

isNonZero :: MPBall -> Bool

CanMinMaxAsymmetric Int MPBall

Associated Types

type MinMaxType Int MPBall :: *

Methods

min :: Int -> MPBall -> MinMaxType Int MPBall

max :: Int -> MPBall -> MinMaxType Int MPBall

CanMinMaxAsymmetric Integer MPBall

Associated Types

type MinMaxType Integer MPBall :: *

Methods

min :: Integer -> MPBall -> MinMaxType Integer MPBall

max :: Integer -> MPBall -> MinMaxType Integer MPBall

CanMinMaxAsymmetric Rational MPBall

Associated Types

type MinMaxType Rational MPBall :: *

Methods

min :: Rational -> MPBall -> MinMaxType Rational MPBall

max :: Rational -> MPBall -> MinMaxType Rational MPBall

CanMinMaxAsymmetric Dyadic MPBall

Associated Types

type MinMaxType Dyadic MPBall :: *

Methods

min :: Dyadic -> MPBall -> MinMaxType Dyadic MPBall

max :: Dyadic -> MPBall -> MinMaxType Dyadic MPBall

CanMinMaxAsymmetric MPBall Int

Associated Types

type MinMaxType MPBall Int :: *

Methods

min :: MPBall -> Int -> MinMaxType MPBall Int

max :: MPBall -> Int -> MinMaxType MPBall Int

CanMinMaxAsymmetric MPBall Integer

Associated Types

type MinMaxType MPBall Integer :: *

Methods

min :: MPBall -> Integer -> MinMaxType MPBall Integer

max :: MPBall -> Integer -> MinMaxType MPBall Integer

CanMinMaxAsymmetric MPBall Rational

Associated Types

type MinMaxType MPBall Rational :: *

Methods

min :: MPBall -> Rational -> MinMaxType MPBall Rational

max :: MPBall -> Rational -> MinMaxType MPBall Rational

CanMinMaxAsymmetric MPBall Dyadic

Associated Types

type MinMaxType MPBall Dyadic :: *

Methods

min :: MPBall -> Dyadic -> MinMaxType MPBall Dyadic

max :: MPBall -> Dyadic -> MinMaxType MPBall Dyadic

CanMinMaxAsymmetric MPBall MPBall

Associated Types

type MinMaxType MPBall MPBall :: *

Methods

min :: MPBall -> MPBall -> MinMaxType MPBall MPBall

max :: MPBall -> MPBall -> MinMaxType MPBall MPBall

HasOrderAsymmetric Int MPBall

Associated Types

type OrderCompareType Int MPBall :: *

Methods

lessThan :: Int -> MPBall -> OrderCompareType Int MPBall

greaterThan :: Int -> MPBall -> OrderCompareType Int MPBall

leq :: Int -> MPBall -> OrderCompareType Int MPBall

geq :: Int -> MPBall -> OrderCompareType Int MPBall

HasOrderAsymmetric Integer MPBall

Associated Types

type OrderCompareType Integer MPBall :: *

Methods

lessThan :: Integer -> MPBall -> OrderCompareType Integer MPBall

greaterThan :: Integer -> MPBall -> OrderCompareType Integer MPBall

leq :: Integer -> MPBall -> OrderCompareType Integer MPBall

geq :: Integer -> MPBall -> OrderCompareType Integer MPBall

HasOrderAsymmetric Rational MPBall

Associated Types

type OrderCompareType Rational MPBall :: *

Methods

lessThan :: Rational -> MPBall -> OrderCompareType Rational MPBall

greaterThan :: Rational -> MPBall -> OrderCompareType Rational MPBall

leq :: Rational -> MPBall -> OrderCompareType Rational MPBall

geq :: Rational -> MPBall -> OrderCompareType Rational MPBall

HasOrderAsymmetric Dyadic MPBall

Associated Types

type OrderCompareType Dyadic MPBall :: *

Methods

lessThan :: Dyadic -> MPBall -> OrderCompareType Dyadic MPBall

greaterThan :: Dyadic -> MPBall -> OrderCompareType Dyadic MPBall

leq :: Dyadic -> MPBall -> OrderCompareType Dyadic MPBall

geq :: Dyadic -> MPBall -> OrderCompareType Dyadic MPBall

HasOrderAsymmetric MPBall Int

Associated Types

type OrderCompareType MPBall Int :: *

Methods

lessThan :: MPBall -> Int -> OrderCompareType MPBall Int

greaterThan :: MPBall -> Int -> OrderCompareType MPBall Int

leq :: MPBall -> Int -> OrderCompareType MPBall Int

geq :: MPBall -> Int -> OrderCompareType MPBall Int

HasOrderAsymmetric MPBall Integer

Associated Types

type OrderCompareType MPBall Integer :: *

Methods

lessThan :: MPBall -> Integer -> OrderCompareType MPBall Integer

greaterThan :: MPBall -> Integer -> OrderCompareType MPBall Integer

leq :: MPBall -> Integer -> OrderCompareType MPBall Integer

geq :: MPBall -> Integer -> OrderCompareType MPBall Integer

HasOrderAsymmetric MPBall Rational

Associated Types

type OrderCompareType MPBall Rational :: *

Methods

lessThan :: MPBall -> Rational -> OrderCompareType MPBall Rational

greaterThan :: MPBall -> Rational -> OrderCompareType MPBall Rational

leq :: MPBall -> Rational -> OrderCompareType MPBall Rational

geq :: MPBall -> Rational -> OrderCompareType MPBall Rational

HasOrderAsymmetric MPBall Dyadic

Associated Types

type OrderCompareType MPBall Dyadic :: *

Methods

lessThan :: MPBall -> Dyadic -> OrderCompareType MPBall Dyadic

greaterThan :: MPBall -> Dyadic -> OrderCompareType MPBall Dyadic

leq :: MPBall -> Dyadic -> OrderCompareType MPBall Dyadic

geq :: MPBall -> Dyadic -> OrderCompareType MPBall Dyadic

HasOrderAsymmetric MPBall MPBall

Associated Types

type OrderCompareType MPBall MPBall :: *

Methods

lessThan :: MPBall -> MPBall -> OrderCompareType MPBall MPBall

greaterThan :: MPBall -> MPBall -> OrderCompareType MPBall MPBall

leq :: MPBall -> MPBall -> OrderCompareType MPBall MPBall

geq :: MPBall -> MPBall -> OrderCompareType MPBall MPBall

HasEqAsymmetric Int MPBall

Associated Types

type EqCompareType Int MPBall :: *

Methods

equalTo :: Int -> MPBall -> EqCompareType Int MPBall

notEqualTo :: Int -> MPBall -> EqCompareType Int MPBall

HasEqAsymmetric Integer MPBall

Associated Types

type EqCompareType Integer MPBall :: *

Methods

equalTo :: Integer -> MPBall -> EqCompareType Integer MPBall

notEqualTo :: Integer -> MPBall -> EqCompareType Integer MPBall

HasEqAsymmetric Rational MPBall

Associated Types

type EqCompareType Rational MPBall :: *

Methods

equalTo :: Rational -> MPBall -> EqCompareType Rational MPBall

notEqualTo :: Rational -> MPBall -> EqCompareType Rational MPBall

HasEqAsymmetric Dyadic MPBall

Associated Types

type EqCompareType Dyadic MPBall :: *

Methods

equalTo :: Dyadic -> MPBall -> EqCompareType Dyadic MPBall

notEqualTo :: Dyadic -> MPBall -> EqCompareType Dyadic MPBall

HasEqAsymmetric MPBall Int

Associated Types

type EqCompareType MPBall Int :: *

Methods

equalTo :: MPBall -> Int -> EqCompareType MPBall Int

notEqualTo :: MPBall -> Int -> EqCompareType MPBall Int

HasEqAsymmetric MPBall Integer

Associated Types

type EqCompareType MPBall Integer :: *

Methods

equalTo :: MPBall -> Integer -> EqCompareType MPBall Integer

notEqualTo :: MPBall -> Integer -> EqCompareType MPBall Integer

HasEqAsymmetric MPBall Rational

Associated Types

type EqCompareType MPBall Rational :: *

Methods

equalTo :: MPBall -> Rational -> EqCompareType MPBall Rational

notEqualTo :: MPBall -> Rational -> EqCompareType MPBall Rational

HasEqAsymmetric MPBall Dyadic

Associated Types

type EqCompareType MPBall Dyadic :: *

Methods

equalTo :: MPBall -> Dyadic -> EqCompareType MPBall Dyadic

notEqualTo :: MPBall -> Dyadic -> EqCompareType MPBall Dyadic

HasEqAsymmetric MPBall MPBall

Associated Types

type EqCompareType MPBall MPBall :: *

Methods

equalTo :: MPBall -> MPBall -> EqCompareType MPBall MPBall

notEqualTo :: MPBall -> MPBall -> EqCompareType MPBall MPBall