AERN2.MP.Ball.Elementary

Description

Elementary operations on arbitrary precision dyadic balls

Ball operations (see also instances)

piBallP :: Precision -> MPBall

Helpers for constructing ball functions

fromApproxWithLipschitz

Arguments

:: (MPFloat -> MPFloat)
fDown: a version of f on MPFloat rounding *downwards*

-> (MPFloat -> MPFloat)
fUp: a version of f on MPFloat rounding *upwards*

-> MPFloat
lip a Lipschitz constant for f, lip > 0

-> MPBall -> MPBall
f on MPBall rounding *outwards*

Computes a real function f from correctly rounded MPFR-approximations and a number lip which is a Lipschitz constant for f, i.e. |f(x) - f(y)| <= lip * |x - y| for all x,y.

Orphan instances

Eq MPBall

Methods

(==) :: MPBall -> MPBall -> Bool

(/=) :: MPBall -> MPBall -> Bool

Floating MPBall

Methods

pi :: MPBall

exp :: MPBall -> MPBall

log :: MPBall -> MPBall

sqrt :: MPBall -> MPBall

(**) :: MPBall -> MPBall -> MPBall

logBase :: MPBall -> MPBall -> MPBall

sin :: MPBall -> MPBall

cos :: MPBall -> MPBall

tan :: MPBall -> MPBall

asin :: MPBall -> MPBall

acos :: MPBall -> MPBall

atan :: MPBall -> MPBall

sinh :: MPBall -> MPBall

cosh :: MPBall -> MPBall

tanh :: MPBall -> MPBall

asinh :: MPBall -> MPBall

acosh :: MPBall -> MPBall

atanh :: MPBall -> MPBall

log1p :: MPBall -> MPBall

expm1 :: MPBall -> MPBall

log1pexp :: MPBall -> MPBall

log1mexp :: MPBall -> MPBall

Fractional MPBall

Methods

(/) :: MPBall -> MPBall -> MPBall

recip :: MPBall -> MPBall

fromRational :: Rational -> MPBall

Num MPBall

Methods

(+) :: MPBall -> MPBall -> MPBall

(-) :: MPBall -> MPBall -> MPBall

(*) :: MPBall -> MPBall -> MPBall

negate :: MPBall -> MPBall

abs :: MPBall -> MPBall

signum :: MPBall -> MPBall

fromInteger :: Integer -> MPBall

Ord MPBall

Methods

compare :: MPBall -> MPBall -> Ordering

(<) :: MPBall -> MPBall -> Bool

(<=) :: MPBall -> MPBall -> Bool

(>) :: MPBall -> MPBall -> Bool

(>=) :: MPBall -> MPBall -> Bool

max :: MPBall -> MPBall -> MPBall

min :: MPBall -> MPBall -> MPBall

CanSqrt MPBall

Associated Types

type SqrtType MPBall :: *

Methods

sqrt :: MPBall -> SqrtType MPBall

CanExp MPBall

Associated Types

type ExpType MPBall :: *

Methods

exp :: MPBall -> ExpType MPBall

CanLog MPBall

Associated Types

type LogType MPBall :: *

Methods

log :: MPBall -> LogType MPBall

CanSinCos MPBall

Associated Types

type SinCosType MPBall :: *

Methods

cos :: MPBall -> SinCosType MPBall

sin :: MPBall -> SinCosType MPBall

CanPow MPBall Rational

Associated Types

type PowType MPBall Rational :: *

Methods

pow :: MPBall -> Rational -> PowType MPBall Rational

CanPow MPBall Dyadic

Associated Types

type PowType MPBall Dyadic :: *

Methods

pow :: MPBall -> Dyadic -> PowType MPBall Dyadic

CanPow MPBall MPBall

Associated Types

type PowType MPBall MPBall :: *

Methods

pow :: MPBall -> MPBall -> PowType MPBall MPBall