AERN2.MP.Float.Tests

Description

Tests for operations on arbitrary precision floats.

To run the tests using stack, execute:

   stack test aern2-mp --test-arguments "-a 1000 -m MPFloat"
   

Documentation

specMPFloat :: Spec

tMPFloat :: T MPFloat

A runtime representative of type MPFloat. Used for specialising polymorphic tests to concrete types.

(=~=) :: MPFloat -> MPFloat -> Property infix 4

approxEqual

Arguments

:: Integer
p precision to guide tolerance

-> MPFloat
LHS of equation

-> MPFloat
RHS of equation

-> Bool

Assert equality of two MPFloat's with tolerance 1/2^p.

approxEqualWithArgs

Arguments

:: [(MPFloat, String)]
intermediate values from which to determine tolerance, their names to report when the equality fails

-> MPFloat
LHS of equation

-> MPFloat
RHS of equation

-> Property

Assert equality of two MPFloat's with tolerance derived from the size and precision of the given intermediate values. When the assertion fails, report the given values using the given names.

frequencyElements :: ConvertibleExactly t Int => [(t, a)] -> Gen a

Orphan instances

Arbitrary MPFloat

Methods

arbitrary :: Gen MPFloat

shrink :: MPFloat -> [MPFloat]