AERN2.MP.UseMPFR.Ball.Type

Description

Arbitrary precision dyadic balls

Auxiliary types

module AERN2.MP.Precision

module AERN2.MP.Accuracy

module AERN2.MP.Enclosure

The Ball type

data MPBall

Constructors

MPBall

Fields

ball_value :: MPFloat

ball_error :: ErrorBound

type CanBeMPBall t = ConvertibleExactly t MPBall

mpBall :: CanBeMPBall t => t -> MPBall

type CanBeMPBallP t = ConvertibleWithPrecision t MPBall

mpBallP :: CanBeMPBallP t => Precision -> t -> MPBall

reducePrecionIfInaccurate :: MPBall -> MPBall

Reduce the precision of the ball centre if the accuracy of the ball is poor.

More precisely, reduce the precision of the centre so that the ulp is approximately (radius / 1024), unless the ulp is already lower than this.

Ball construction/extraction functions

endpointsMP :: MPBall -> (MPFloat, MPFloat)

fromEndpointsMP :: MPFloat -> MPFloat -> MPBall