AERN2.MP.UseMPFR.Float.Conversions

Description

Conversions and comparisons of arbitrary precision floating point numbers

Currently, we use hmpfr when compiling with ghc 7.10 and higher and haskell-mpfr when compiling with ghc 7.8.

MPFloat to other types (see also instances)

toDoubleUp :: MPFloat -> Double

toDoubleDown :: MPFloat -> Double

MPFloat constructors (see also instances)

type CanBeMPFloat t = ConvertibleExactly t MPFloat

mpFloat :: CanBeMPFloat t => t -> MPFloat

fromIntegerUp :: Precision -> Integer -> MPFloat

fromIntegerDown :: Precision -> Integer -> MPFloat

fromRationalUp :: Precision -> Rational -> MPFloat

fromRationalDown :: Precision -> Rational -> MPFloat

Orphan instances

CanRound MPFloat

Methods

properFraction :: MPFloat -> (Integer, MPFloat)

truncate :: MPFloat -> Integer

round :: MPFloat -> Integer

ceiling :: MPFloat -> Integer

floor :: MPFloat -> Integer

CanTestPosNeg MPFloat

Methods

isCertainlyPositive :: MPFloat -> Bool

isCertainlyNonNegative :: MPFloat -> Bool

isCertainlyNegative :: MPFloat -> Bool

isCertainlyNonPositive :: MPFloat -> Bool

CanTestZero MPFloat

Methods

isCertainlyZero :: MPFloat -> Bool

isCertainlyNonZero :: MPFloat -> Bool

HasNorm MPFloat

Methods

getNormLog :: MPFloat -> NormLog

Convertible MPFloat Double

Methods

safeConvert :: MPFloat -> ConvertResult Double

CanMinMaxAsymmetric MPFloat MPFloat

Associated Types

type MinMaxType MPFloat MPFloat :: *

Methods

min :: MPFloat -> MPFloat -> MinMaxType MPFloat MPFloat

max :: MPFloat -> MPFloat -> MinMaxType MPFloat MPFloat

HasOrderAsymmetric Int MPFloat

Associated Types

type OrderCompareType Int MPFloat :: *

Methods

lessThan :: Int -> MPFloat -> OrderCompareType Int MPFloat

greaterThan :: Int -> MPFloat -> OrderCompareType Int MPFloat

leq :: Int -> MPFloat -> OrderCompareType Int MPFloat

geq :: Int -> MPFloat -> OrderCompareType Int MPFloat

HasOrderAsymmetric Integer MPFloat

Associated Types

type OrderCompareType Integer MPFloat :: *

Methods

lessThan :: Integer -> MPFloat -> OrderCompareType Integer MPFloat

greaterThan :: Integer -> MPFloat -> OrderCompareType Integer MPFloat

leq :: Integer -> MPFloat -> OrderCompareType Integer MPFloat

geq :: Integer -> MPFloat -> OrderCompareType Integer MPFloat

HasOrderAsymmetric Rational MPFloat

Associated Types

type OrderCompareType Rational MPFloat :: *

Methods

lessThan :: Rational -> MPFloat -> OrderCompareType Rational MPFloat

greaterThan :: Rational -> MPFloat -> OrderCompareType Rational MPFloat

leq :: Rational -> MPFloat -> OrderCompareType Rational MPFloat

geq :: Rational -> MPFloat -> OrderCompareType Rational MPFloat

HasOrderAsymmetric MPFloat Int

Associated Types

type OrderCompareType MPFloat Int :: *

Methods

lessThan :: MPFloat -> Int -> OrderCompareType MPFloat Int

greaterThan :: MPFloat -> Int -> OrderCompareType MPFloat Int

leq :: MPFloat -> Int -> OrderCompareType MPFloat Int

geq :: MPFloat -> Int -> OrderCompareType MPFloat Int

HasOrderAsymmetric MPFloat Integer

Associated Types

type OrderCompareType MPFloat Integer :: *

Methods

lessThan :: MPFloat -> Integer -> OrderCompareType MPFloat Integer

greaterThan :: MPFloat -> Integer -> OrderCompareType MPFloat Integer

leq :: MPFloat -> Integer -> OrderCompareType MPFloat Integer

geq :: MPFloat -> Integer -> OrderCompareType MPFloat Integer

HasOrderAsymmetric MPFloat Rational

Associated Types

type OrderCompareType MPFloat Rational :: *

Methods

lessThan :: MPFloat -> Rational -> OrderCompareType MPFloat Rational

greaterThan :: MPFloat -> Rational -> OrderCompareType MPFloat Rational

leq :: MPFloat -> Rational -> OrderCompareType MPFloat Rational

geq :: MPFloat -> Rational -> OrderCompareType MPFloat Rational

HasOrderAsymmetric MPFloat MPFloat

Associated Types

type OrderCompareType MPFloat MPFloat :: *

Methods

lessThan :: MPFloat -> MPFloat -> OrderCompareType MPFloat MPFloat

greaterThan :: MPFloat -> MPFloat -> OrderCompareType MPFloat MPFloat

leq :: MPFloat -> MPFloat -> OrderCompareType MPFloat MPFloat

geq :: MPFloat -> MPFloat -> OrderCompareType MPFloat MPFloat

HasEqAsymmetric Int MPFloat

Associated Types

type EqCompareType Int MPFloat :: *

Methods

equalTo :: Int -> MPFloat -> EqCompareType Int MPFloat

notEqualTo :: Int -> MPFloat -> EqCompareType Int MPFloat

HasEqAsymmetric Integer MPFloat

Associated Types

type EqCompareType Integer MPFloat :: *

Methods

equalTo :: Integer -> MPFloat -> EqCompareType Integer MPFloat

notEqualTo :: Integer -> MPFloat -> EqCompareType Integer MPFloat

HasEqAsymmetric Rational MPFloat

Associated Types

type EqCompareType Rational MPFloat :: *

Methods

equalTo :: Rational -> MPFloat -> EqCompareType Rational MPFloat

notEqualTo :: Rational -> MPFloat -> EqCompareType Rational MPFloat

HasEqAsymmetric MPFloat Int

Associated Types

type EqCompareType MPFloat Int :: *

Methods

equalTo :: MPFloat -> Int -> EqCompareType MPFloat Int

notEqualTo :: MPFloat -> Int -> EqCompareType MPFloat Int

HasEqAsymmetric MPFloat Integer

Associated Types

type EqCompareType MPFloat Integer :: *

Methods

equalTo :: MPFloat -> Integer -> EqCompareType MPFloat Integer

notEqualTo :: MPFloat -> Integer -> EqCompareType MPFloat Integer

HasEqAsymmetric MPFloat Rational

Associated Types

type EqCompareType MPFloat Rational :: *

Methods

equalTo :: MPFloat -> Rational -> EqCompareType MPFloat Rational

notEqualTo :: MPFloat -> Rational -> EqCompareType MPFloat Rational

HasEqAsymmetric MPFloat MPFloat

Associated Types

type EqCompareType MPFloat MPFloat :: *

Methods

equalTo :: MPFloat -> MPFloat -> EqCompareType MPFloat MPFloat

notEqualTo :: MPFloat -> MPFloat -> EqCompareType MPFloat MPFloat

ConvertibleExactly Int MPFloat

Methods

safeConvertExactly :: Int -> ConvertResult MPFloat

ConvertibleExactly Integer MPFloat

Methods

safeConvertExactly :: Integer -> ConvertResult MPFloat

ConvertibleExactly MPFloat Rational

Methods

safeConvertExactly :: MPFloat -> ConvertResult Rational