AERN2.MP.Ball.Elementary

Description

Elementary operations on arbitrary precision dyadic balls

Ball operations (see also instances)

piBallP :: Precision -> MPBall

Helpers for constructing ball functions

fromApproxWithLipschitz

Arguments

:: (MPFloat -> MPFloat)
fDown: a version of f on MPFloat rounding *downwards*

-> (MPFloat -> MPFloat)
fUp: a version of f on MPFloat rounding *upwards*

-> MPFloat
lip a Lipschitz constant for f, lip > 0

-> MPBall -> MPBall
f on MPBall rounding *outwards*

Computes a real function f from correctly rounded MPFR-approximations and a number lip which is a Lipschitz constant for f, i.e. |f(x) - f(y)| <= lip * |x - y| for all x,y.

Orphan instances

CanSqrt MPBall

Associated Types

type SqrtType MPBall :: *

Methods

sqrt :: MPBall -> SqrtType MPBall

CanExp MPBall

Associated Types

type ExpType MPBall :: *

Methods

exp :: MPBall -> ExpType MPBall

CanLog MPBall

Associated Types

type LogType MPBall :: *

Methods

log :: MPBall -> LogType MPBall

CanSinCos MPBall

Associated Types

type SinCosType MPBall :: *

Methods

cos :: MPBall -> SinCosType MPBall

sin :: MPBall -> SinCosType MPBall

CanPow MPBall Rational

Associated Types

type PowTypeNoCN MPBall Rational :: *

type PowType MPBall Rational :: *

Methods

powNoCN :: MPBall -> Rational -> PowTypeNoCN MPBall Rational

pow :: MPBall -> Rational -> PowType MPBall Rational

CanPow MPBall Dyadic

Associated Types

type PowTypeNoCN MPBall Dyadic :: *

type PowType MPBall Dyadic :: *

Methods

powNoCN :: MPBall -> Dyadic -> PowTypeNoCN MPBall Dyadic

pow :: MPBall -> Dyadic -> PowType MPBall Dyadic

CanPow MPBall MPBall

Associated Types

type PowTypeNoCN MPBall MPBall :: *

type PowType MPBall MPBall :: *

Methods

powNoCN :: MPBall -> MPBall -> PowTypeNoCN MPBall MPBall

pow :: MPBall -> MPBall -> PowType MPBall MPBall