AERN2.MP.Enclosure

Description

Enclosure classes and operations.

Documentation

class IsBall t where

Minimal complete definition

centre, centreAsBallAndRadius, updateRadius

Associated Types

type CentreType t

Methods

centre :: t -> CentreType t

centreAsBallAndRadius :: t -> (t, ErrorBound)

centreAsBall :: t -> t

radius :: t -> ErrorBound

updateRadius :: (ErrorBound -> ErrorBound) -> t -> t

makeExactCentre :: IsBall t => t -> t

When the radius of the ball is implicitly contributed to by imprecision in the centre (eg if the centre is a polynomial with inexact coefficients), move all that imprecision to the explicit radius, making the centre exact. This may lose some information, but as a ball is equivalent to the original. For MPBall this function is pointless because it is equivalent to the identity.

class IsInterval i e where

Minimal complete definition

fromEndpoints, endpoints

Methods

fromEndpoints :: e -> e -> i

endpoints :: i -> (e, e)

intervalFunctionByEndpoints

Arguments

:: (IsInterval t t, HasEqCertainly t t)

=> (t -> t)
fThin: a version of f that works well on thin intervals

-> t -> t
f on *large* intervals

Computes a *monotone* ball function f on intervals using the interval endpoints.

intervalFunctionByEndpointsUpDown

Arguments

:: IsInterval t e

=> (e -> e)
fDown: a version of f working on endpoints, rounded down

-> (e -> e)
fUp: a version of f working on endpoints, rounded up

-> t -> t
f on intervals rounding *outwards*

Computes a *monotone* ball function f on intervals using the interval endpoints.

class CanTestContains dom e where

Minimal complete definition

contains

Methods

contains :: dom -> e -> Bool

Test if e is inside dom.

class CanMapInside dom e where

Minimal complete definition

mapInside

Methods

mapInside :: dom -> e -> e

Return some value contained in dom. The returned value does not have to equal the given e even if e is already inside dom. All elements of dom should be covered with roughly the same probability when calling this function for evenly distributed e's.

This function is intended mainly for generating values inside dom for randomised tests.

specCanMapInside :: (CanMapInside d e, CanTestContains d e, Arbitrary d, Arbitrary e, Show d, Show e) => T d -> T e -> Spec

class CanIntersectAsymmetric e1 e2 where

Minimal complete definition

intersect

Associated Types

type IntersectionType e1 e2

Methods

intersect :: e1 -> e2 -> IntersectionType e1 e2

type CanIntersect e1 e2 = (CanIntersectAsymmetric e1 e2, CanIntersectAsymmetric e1 e2, IntersectionType e1 e2 ~ IntersectionType e2 e1)

type CanIntersectCNBy e1 e2 = (CanIntersect e1 e2, IntersectionType e1 e2 ~ EnsureCN e1, CanIntersect (EnsureCN e1) e2, IntersectionType (EnsureCN e1) e2 ~ EnsureCN e1)

type CanIntersectCNSameType e1 = (CanIntersectCNBy e1 e1, CanIntersect (EnsureCN e1) (EnsureCN e1), IntersectionType (EnsureCN e1) (EnsureCN e1) ~ EnsureCN e1)

class CanUnionAsymmetric e1 e2 where

Minimal complete definition

union

Associated Types

type UnionType e1 e2

Methods

union :: e1 -> e2 -> UnionType e1 e2

type CanUnion e1 e2 = (CanUnionAsymmetric e1 e2, CanUnionAsymmetric e1 e2, UnionType e1 e2 ~ UnionType e2 e1)

type CanUnionCNBy e1 e2 = (CanUnion e1 e2, UnionType e1 e2 ~ EnsureCN e1, CanUnion (EnsureCN e1) e2, UnionType (EnsureCN e1) e2 ~ EnsureCN e1)

type CanUnionCNSameType e1 = (CanUnionCNBy e1 e1, CanUnion (EnsureCN e1) (EnsureCN e1), UnionType (EnsureCN e1) (EnsureCN e1) ~ EnsureCN e1)

Orphan instances

(CanUnionCNSameType t, CanEnsureCN t) => HasIfThenElse (Maybe Bool) t

Associated Types

type IfThenElseType (Maybe Bool) t :: *

Methods

ifThenElse :: Maybe Bool -> t -> t -> IfThenElseType (Maybe Bool) t